Add user logs

+++ routes/api/v1/handlers/album_handler.go
@@ -1,6 +1,9 @@
 package handlers
 
 import (
+	"errors"
+	"fmt"
+	"log"
 	"net/http"
 
 	"GoAlbums/internal/service"
@@ -45,17 +48,39 @@ func CreateAlbum(c *gin.Context) {
 		return
 	}
 
+	userID, exists := c.Get("userID")
+	if !exists {
+		errorResponse, statusCode := helpers.CustomError(errors.New("userID not found"))
+		c.JSON(statusCode, errorResponse)
+		return
+	}
+
+	newAlbum.UpdatedBy = userID.(string)
+
 	album, err := service.CreateAlbum(newAlbum)
 	if err != nil {
 		errorResponse, statusCode := helpers.CustomError(err)
 		c.JSON(statusCode, errorResponse)
 		return
 	}
+
+	userLogs := service.UserLogs{
+		UserId:   userID.(string),
+		Activity: fmt.Sprintf("Created album -> %v", album),
+	}
+
+	if _, err := service.CreateUserLogs(userLogs); err != nil {
+		log.Printf("Failed to log user activity: %v", err)
+	}
+
 	c.JSON(http.StatusCreated, album)
 }
 
 func UpdateAlbum(c *gin.Context) {
+
 	id := c.Param("id")
+	oldAlbum, _ := service.GetAlbumById(id)
+
 	var updatedAlbum service.Album
 	if err := c.BindJSON(&updatedAlbum); err != nil {
 		errorResponse, statusCode := helpers.CustomError(err)
@@ -69,12 +94,31 @@ func UpdateAlbum(c *gin.Context) {
 		return
 	}
 
+	userID, exists := c.Get("userID")
+	if !exists {
+		errorResponse, statusCode := helpers.CustomError(errors.New("userID not found"))
+		c.JSON(statusCode, errorResponse)
+		return
+	}
+
+	updatedAlbum.UpdatedBy = userID.(string)
+
 	album, err := service.UpdateAlbum(id, updatedAlbum)
 	if err != nil {
 		errorResponse, statusCode := helpers.CustomError(err)
 		c.JSON(statusCode, errorResponse)
 		return
 	}
+
+	userLogs := service.UserLogs{
+		UserId:   userID.(string),
+		Activity: fmt.Sprintf("Modified album -> %v > %v", oldAlbum, album),
+	}
+
+	if _, err := service.CreateUserLogs(userLogs); err != nil {
+		log.Printf("Failed to log user activity: %v", err)
+	}
+
 	c.JSON(http.StatusOK, album)
 }
 
@@ -86,11 +130,30 @@ func DeleteAlbums(c *gin.Context) {
 		return
 	}
 
+	userID, exists := c.Get("userID")
+	if !exists {
+		errorResponse, statusCode := helpers.CustomError(errors.New("userID not found"))
+		c.JSON(statusCode, errorResponse)
+		return
+	}
+
 	err := service.DeleteAlbums(ids)
 	if err != nil {
 		errorResponse, statusCode := helpers.CustomError(err)
 		c.JSON(statusCode, errorResponse)
 		return
 	}
+
+	for _, id := range ids {
+		userLog := service.UserLogs{
+			UserId:   userID.(string),
+			Activity: fmt.Sprintf("Deleted album -> %s", id),
+		}
+
+		if _, err := service.CreateUserLogs(userLog); err != nil {
+			log.Printf("Failed to log user activity for album ID %s: %v", id, err)
+		}
+	}
+
 	c.JSON(http.StatusOK, gin.H{"message": "Album(s) deleted"})
 }
